+++ pywatts/modules/__init__.py
@@ -23,3 +23,4 @@ from pywatts.modules.preprocessing.resampler import Resampler
 from pywatts.modules.preprocessing.sampler import Sampler
 from pywatts.modules.preprocessing.slicer import Slicer
 from pywatts.modules.wrappers import *
+from pywatts.modules.feature_extraction.statistics_extraction import StatisticExtraction, StatisticFeature