+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 0.3.1 Unreleased
+
+### Added
+* Add alpha to rolling mean ([#225](https://github.com/KIT-IAI/pyWATTS/issues/225))
+
 
 ## 0.3.0 - 2022-31-03